Move remaining builder classes into the FromFedora module
